各種クラスのweak_ptrの別名定義をまとめるヘッダファイルを追加。 ※個々のヘッダファイルに定義していると循環参照の原因になったりして面倒なので。
