+++ Server/AnpanMMOServer/AnpanMMOServer.vcxproj
@@ -202,6 +202,7 @@
     <ClInclude Include="TickManager.h" />
     <ClInclude Include="TypeDefs.h" />
     <ClInclude Include="Util.h" />
+    <ClInclude Include="WeakPtrDefine.h" />
     <ClInclude Include="WordCheckServer\WordCheckPacketReceiver.h" />
     <ClInclude Include="WordCheckServer\WordCheckServerConnection.h" />
   </ItemGroup>
