+++ Server/AnpanMMOServer/AnpanMMOServer.vcxproj.filters
@@ -255,6 +255,9 @@
     <ClInclude Include="Party\PartyManager.h">
       <Filter>ヘッダー ファイル</Filter>
     </ClInclude>
+    <ClInclude Include="WeakPtrDefine.h">
+      <Filter>ヘッダー ファイル</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="MemoryBuffer.cpp">
